Initial proof-of-concept of the hybrid fiji.exe generator
It is a bit annoying that there are fiji-win32.exe, fiji-win64.exe,
fiji-linux and fiji-linux64 as well as the executables in Contents/MacOS/
in an unpacked fiji-all.zip.

The problem is of course that the launchers need to be platform and
architecture specific.

So here is the idea: make a regular, 32-bit Windows program (so that
Windows can execute it) and then hack it a bit so that it is a valid
shell script, too (so that everything else can execute it). In both
cases, determine the actual platform and architecture and then hand
off to the matching ImageJ launcher.

This is just a proof-of-concept: it only shows how to make a hybrid
fiji.exe, but the program does not yet do what it is supposed to do.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>