Do not hard-code size of the PE part
When executing the script, we need to skip the PE part of the hybrid
launcher (an earlier attempt to ignore it via 'cat << \EOFEOFEF'
failed because the single quotes and back ticks in that section were
still not ignored by the shell).

The previous approach hard-coded the size of the PE part, but when
attaching an icon to the .exe, that size changes.

Therefore, let's use a unique marker and ask sed to skip the PE part
dynamically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>